Fix slack notification failure (#485)
Issue: slack notification failed:
https://github.com/rgovindji/modclub/actions/runs/6614930509/job/17965966317
Suspect it is due to the oversize of slack text msg.

Fix: set a limit on the length.

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Implemented a solution to prevent Slack notification failures
during deployment. The update ensures that the content of the
`WHATS_NEW_SINCE_LAST_DEP` environment variable does not exceed the
Slack message character limit. If the message is too long, it will be
truncated and appended with ellipsis, ensuring successful and consistent
notifications about new changes since the last deployment.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->